feat: 회원가입, 로그인 부분 추가, security 적용, 프로젝트 구조에 맞게 클래스 위치 변경, service, repository 부분 수정

+++ src/main/java/com/project/eatTogether/application/dto/RsRestaurantDTO.java
@@ -20,6 +20,6 @@ public class RsRestaurantDTO {
     private String rsState; // 식당 상태 (오픈, 마감)
     private int rsReviewCount; // 식당 리뷰 수
     private int rsBookmarkCount; // 식당 북마크 수
-    private int rsAvgRate; // 식당 평균 평점
+    private Double rsAvgRate; // 식당 평균 평점
     private int rsReservationCount; // 식당 예약 횟수
 }
